Apply changes after review
