+++ saleor/graphql/payment/types.py
@@ -156,10 +156,14 @@ class Payment(CountableDjangoObjectType):
 
 class PaymentInitialized(graphene.ObjectType):
     class Meta:
-        description = "Represents a initialized data by payment gateway."
+        description = (
+            "Server-side data generated by a payment gateway. Optional step when the "
+            "payment provider requires an additional action to initialize payment "
+            "session."
+        )
 
-    gateway = graphene.String(description="Payment gateway.", required=True)
+    gateway = graphene.String(description="ID of a payment gateway.", required=True)
     name = graphene.String(description="Payment gateway name.", required=True)
     data = graphene.JSONString(
-        description="Initiliazed data by gateway.", required=False
+        description="Initialized data by gateway.", required=False
     )
