+++ saleor/plugins/base_plugin.py
@@ -34,14 +34,14 @@ class ConfigurationTypeField:
     STRING = "String"
     BOOLEAN = "Boolean"
     SECRET = "Secret"
-    SECRET_TEXT = "SecretText"
+    SECRET_MULTILINE = "SecretText"
     PASSWORD = "Password"
     CHOICES = [
         (STRING, "Field is a String"),
         (BOOLEAN, "Field is a Boolean"),
         (SECRET, "Field is a Secret"),
         (PASSWORD, "Field is a Password"),
-        (SECRET_TEXT, "Field is a Secret multiline"),
+        (SECRET_MULTILINE, "Field is a Secret multiline"),
     ]
